Add Jest testing framework
I've added the Jest testing framework to the project to enable automated testing of our codebase. Jest is a popular JavaScript testing library that will help us ensure the reliability and correctness of our code. With Jest in place, we can write and run tests for various parts of our application, including MongoDB operations.