Add favicon file and enable functionality for showChart element
With this commit, a favicon file has been added to the root directory of the website, and JavaScript code to trigger the functionality when clicking the showChart element has been implemented. Now, the website can display a favicon, and users can click the "Show Chart" button to view the budget chart.