Resize canvas on button click
This commit updates the canvas size to a new width and height when the "Show Chart" button is clicked. The JavaScript event listener listens for clicks on the "Show Chart" button and then sets the canvas width and height to the specified values (in this case, 300 pixels width and 200 pixels height). Additionally, it triggers the function createGraphs() to generate graphs with the new dimensions.

This change allows for a more compact display of the chart when the button is clicked.